Move input model package
