+++ src/main/java/org/bric/gui/swing/ArrayListTransferHandler.java
@@ -4,7 +4,7 @@
  */
 package org.bric.gui.swing;
 
-import org.bric.core.model.ImportedImage;
+import org.bric.core.input.model.ImportedImage;
 
 import javax.swing.*;
 import java.awt.datatransfer.DataFlavor;
