+++ src/main/java/org/bric/processor/FileNameService.java
@@ -1,6 +1,6 @@
 package org.bric.processor;
 
-import org.bric.core.model.ImportedImage;
+import org.bric.core.input.model.ImportedImage;
 import org.bric.core.model.output.OutputType;
 import org.bric.utils.Utils;
