+++ src/main/java/org/bric/processor/ImageProcessHandler.java
@@ -7,8 +7,8 @@ import com.itextpdf.text.Rectangle;
 import com.itextpdf.text.pdf.PdfWriter;
 import org.apache.sanselan.ImageFormat;
 import org.apache.sanselan.Sanselan;
-import org.bric.core.model.ImportedImage;
-import org.bric.core.model.input.InputType;
+import org.bric.core.input.model.ImportedImage;
+import org.bric.core.input.model.InputType;
 import org.bric.core.model.output.OutputParameters;
 import org.bric.core.model.output.OutputType;
 import org.bric.gui.inputOutput.ProgressBarFrame;
